glom worker into runner
