+++ python/cog/server/worker.py
@@ -1,9 +1,7 @@
 import asyncio
 import contextlib
 import inspect
-import logging
 import multiprocessing
-import os
 import signal
 import sys
 import traceback
@@ -35,7 +33,6 @@ from .eventtypes import (
 from .exceptions import (
     CancelationException,
     FatalWorkerException,
-    InvalidStateException,
 )
 from .helpers import AsyncPipe, StreamRedirector, WrappedStream, race
 
@@ -88,205 +85,6 @@ class Mux:
                 break
         if self.fatal:
             raise self.fatal
-
-
-class Worker:
-    def __init__(
-        self, predictor_ref: str, tee_output: bool = True, concurrency: int = 1
-    ) -> None:
-        self._state = WorkerState.NEW
-        self._semaphore = asyncio.Semaphore(concurrency)
-        self._concurrency = concurrency
-
-        # A pipe with which to communicate with the child worker.
-        events, child_events = _spawn.Pipe()
-        self._child = _ChildWorker(predictor_ref, child_events, tee_output)
-        self._events: "AsyncPipe[tuple[str, PublicEventType]]" = AsyncPipe(
-            events, self._child.is_alive
-        )
-        # shutdown requested
-        self._shutting_down = False
-        # stop reading events
-        self._terminating = asyncio.Event()
-        self._mux = Mux(self._terminating)
-        self._predictions_in_flight = set()
-
-    def setup_is_allowed(self) -> bool:
-        return self._state == WorkerState.NEW
-
-    def setup(self) -> AsyncIterator[PublicEventType]:
-        self._assert_state(WorkerState.NEW)
-        self._state = WorkerState.STARTING
-
-        async def inner() -> AsyncIterator[PublicEventType]:
-            # in 3.10 Event started doing get_running_loop
-            # previously it stored the loop when created, which causes an error in tests
-            if sys.version_info < (3, 10):
-                self._terminating = self._mux.terminating = asyncio.Event()
-
-            self._child.start()
-            self._ensure_event_reader()
-            async for event in self._mux.read("SETUP", poll=0.1):
-                yield event
-                if isinstance(event, Done):
-                    if event.error:
-                        raise FatalWorkerException(
-                            "Predictor errored during setup: " + event.error_detail
-                        )
-                    self._state = WorkerState.IDLE
-
-        return inner()
-
-    def state_from_predictions_in_flight(self) -> WorkerState:
-        valid_states = {WorkerState.IDLE, WorkerState.PROCESSING, WorkerState.BUSY}
-        if self._state not in valid_states:
-            raise InvalidStateException(
-                f"Invalid operation: state is {self._state} (must be IDLE, PROCESSING, or BUSY)"
-            )
-        if len(self._predictions_in_flight) == self._concurrency:
-            return WorkerState.BUSY
-        if len(self._predictions_in_flight) == 0:
-            return WorkerState.IDLE
-        return WorkerState.PROCESSING
-
-    def is_busy(self) -> bool:
-        return self._state not in {WorkerState.PROCESSING, WorkerState.IDLE}
-
-    def enter_predict(self, id: str) -> None:
-        if self.is_busy():
-            raise InvalidStateException(
-                f"Invalid operation: state is {self._state} (must be processing or idle)"
-            )
-        if self._shutting_down:
-            raise InvalidStateException(
-                "cannot accept new predictions because shutdown requested"
-            )
-        self._predictions_in_flight.add(id)
-        self._state = self.state_from_predictions_in_flight()
-
-    def exit_predict(self, id: str) -> None:
-        self._predictions_in_flight.remove(id)
-        self._state = self.state_from_predictions_in_flight()
-
-    @contextlib.contextmanager
-    def prediction_ctx(self, id: str) -> Iterator[None]:
-        self.enter_predict(id)
-        try:
-            yield
-        finally:
-            self.exit_predict(id)
-
-    async def inner_async_predict(
-        self, input: PredictionInput, poll: Optional[float]
-    ) -> AsyncIterator[PublicEventType]:
-        # perhaps we can download the files and mutate input around here
-        # we're already tracking allowed concurrency with _predictions_in_flight
-        # but we're also doing a semaphore, just in case
-        # the semaphore also allows us to tell sync-submitted-work vs actually started async work
-        async with self._semaphore:
-            self._events.send(input)
-            print("worker sent", input)
-            async for e in self._mux.read(input.id, poll=poll):
-                yield e
-        # we may still need to upload files but we could probably already accept new work
-        # it might be nice to exit here already
-        # self.exit_predict(input.id)
-
-    @contextlib.contextmanager
-    def good_predict(  # better
-        self, input: PredictionInput, poll: Optional[float] = None
-    ) -> Iterator[AsyncIterator[PublicEventType]]:
-        with self.prediction_ctx(input.id):
-            yield self.inner_async_predict(input, poll)
-
-    def predict(  # very bad
-        self, input: PredictionInput, poll: Optional[float] = None
-    ) -> AsyncIterator[PublicEventType]:
-        if isinstance(input, dict):
-            input = PredictionInput(payload=input, id="1")  # just for tests
-        # still not right for tests
-        # this means the prediction is marked as done even before the generator is entered
-        with self.prediction_ctx(input.id):
-            return self.inner_async_predict(input, poll)
-
-    def shutdown(self) -> None:
-        if self._state == WorkerState.DEFUNCT:
-            return
-        # shutdown requested, but keep reading events
-        self._shutting_down = True
-
-        if self._child.is_alive():
-            self._events.send(Shutdown())
-
-    def terminate(self) -> None:
-        if self._state == WorkerState.DEFUNCT:
-            return
-
-        self._terminating.set()
-        self._state = WorkerState.DEFUNCT
-
-        if self._child.is_alive():
-            self._child.terminate()
-            self._child.join()
-        self._events.shutdown()
-        if self._read_events_task:
-            self._read_events_task.cancel()
-
-    # FIXME: this will need to use a combination
-    # of signals and Cancel events on the pipe
-    def cancel(self, id: str) -> None:
-        if id not in self._predictions_in_flight:
-            print("id not there", id, self._predictions_in_flight)
-            raise KeyError
-        if self._child.is_alive() and self._child.pid is not None:
-            os.kill(self._child.pid, signal.SIGUSR1)
-            print("sent cancel")
-            self._events.send(Cancel(id))
-            # this should probably check self._semaphore._value == self._concurrent
-
-    def _assert_state(self, state: WorkerState) -> None:
-        if self._state != state:
-            raise InvalidStateException(
-                f"Invalid operation: state is {self._state} (must be {state})"
-            )
-
-    _read_events_task: "Optional[asyncio.Task[None]]" = None
-
-    def _ensure_event_reader(self) -> None:
-        def handle_error(task: "asyncio.Task[None]") -> None:
-            if task.cancelled():
-                return
-            exc = task.exception()
-            if exc:
-                logging.error("caught exception", exc_info=exc)
-
-        if not self._read_events_task:
-            self._read_events_task = asyncio.create_task(self._read_events())
-            self._read_events_task.add_done_callback(handle_error)
-
-    async def _read_events(self) -> None:
-        while self._child.is_alive() and not self._terminating.is_set():
-            # this can still be running when the task is destroyed
-            result = await self._events.coro_recv_with_exit(self._terminating)
-            print("reader got", result)
-            if result is None:  # event loop closed or child died
-                break
-            id, event = result
-            if id == "LOG" and self._state == WorkerState.STARTING:
-                id = "SETUP"
-            if id == "LOG" and len(self._predictions_in_flight) == 1:
-                id = list(self._predictions_in_flight)[0]
-            await self._mux.write(id, event)
-        # If we dropped off the end off the end of the loop, check if it's
-        # because the child process died.
-        if not self._child.is_alive() and not self._terminating.is_set():
-            exitcode = self._child.exitcode
-            self._mux.fatal = FatalWorkerException(
-                f"Prediction failed for an unknown reason. It might have run out of memory? (exitcode {exitcode})"
-            )
-        # this is the same event as _terminating
-        # we need to set it so mux.reads wake up and throw an error if needed
-        self._mux.terminating.set()
 
 
 class _ChildWorker(_spawn.Process):  # type: ignore
